% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMoodtracker.R
\name{addMoodtracker}
\alias{addMoodtracker}
\title{function combining data from introductiory questionnaire with data from moodtracker
Returns data frame where one row == one time moodtracker was completed}
\usage{
addMoodtracker(MOOD, subDatabase)
}
\arguments{
\item{MOOD}{data frame containing moodtracker data}

\item{subDatabase}{data base with INTRO questionnaire data}
}
\description{
function combining data from introductiory questionnaire with data from moodtracker
Returns data frame where one row == one time moodtracker was completed
}
