% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSummaryVariables.R
\name{addSummaryVariables}
\alias{addSummaryVariables}
\title{function adding any aggragated variables of interest}
\usage{
addSummaryVariables(data)
}
\arguments{
\item{data}{data set with subject, moodtracker, task and questinnaire data}
}
\description{
function adding any aggragated variables of interest
}
