% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaskData.R
\name{addTaskData}
\alias{addTaskData}
\title{Function for adding Questionnaire data
--> assumes moodtracer and subject data has already been combined}
\usage{
addTaskData(data, subDatabase, taskID, taskData, Q = F)
}
\arguments{
\item{data}{Preprocessed data frame that questionnaire data shuld be added to}

\item{subDatabase}{Data frame containing subject information. Used to match data and subject}

\item{taskID}{name of task/questionnaire that should be added}

\item{taskData}{data frame containing task/questionnare data}

\item{Q}{add questionnaire data? set to true if yes}
}
\description{
Function for adding Questionnaire data
--> assumes moodtracer and subject data has already been combined
}
