% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDatabase.R
\name{createDatabase}
\alias{createDatabase}
\title{Function for creating dataframe containing unique subject information-
Used as reference to match participants data across frames}
\usage{
createDatabase(INTRO)
}
\arguments{
\item{INTRO}{data frame containing data from introductory questionnaire}
}
\description{
Function for creating dataframe containing unique subject information-
Used as reference to match participants data across frames
}
