% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSubjectTime.R
\name{matchSubID}
\alias{matchSubID}
\title{Function for matching subject data with information stored in subject database
Returns subI  unique subID from database}
\usage{
matchSubID(subINFO, subDatabase)
}
\arguments{
\item{subINFO}{c("study_id", "deviceModel", "deviceUUID") of subject we want to match}

\item{subDatabase}{database containing subject data for each subject}
}
\description{
Function for matching subject data with information stored in subject database
Returns subI  unique subID from database
}
