% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSubjectTime.R
\name{matchTimepoint}
\alias{matchTimepoint}
\title{function matching timepoint of task data with moodtracker}
\usage{
matchTimepoint(data, tp, subID, taskData)
}
\arguments{
\item{data}{data with subject info and moodtracker (+ task/questionnaie data)}

\item{tp}{timepoint (row number) in task/questionnaire data}

\item{subID}{subject identifiication (returned from matchSubID)}

\item{taskData}{task/questionnaire data}
}
\description{
function matching timepoint of task data with moodtracker
}
