% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{PreProcessing}
\alias{PreProcessing}
\title{Main Preprocessing function
--> calls other sub-functions and runs them in the appropriate order}
\usage{
PreProcessing(path, pathOUT, return = T)
}
\arguments{
\item{path}{directory with data files}

\item{pathOUT}{directory that data should be saved to}

\item{return}{TRUE/FALSE, return data to R environment?}
}
\value{
data --> preprocessed/combined data frame
}
\description{
Main Preprocessing function
--> calls other sub-functions and runs them in the appropriate order
}
